Rename theme file to avoid routing
